volesilla_utils: check db schema
label: feature

There is new table called internal in our DB schema. This table
contains information about used database schema version and timestamp
when was database upgraded to this schema version.

Additionaly, this patch adds new command check to volesilla_utils which
compares schema version in volesilla application and in the given
database file.

This patch is first step to updating scripts for database schema which
we need for enterprise deployment.

Resolves #43